kernel-module-split: ignore missing dirs
fixes

     0199:    for dir in ["%s/etc/modprobe.d" % (dvar), "%s/etc/modules-load.d" % (dvar), "%s/etc" % (dvar)]:
 *** 0200:        if len(os.listdir(dir)) == 0:
     0201:            os.rmdir(dir)

Exception: FileNotFoundError: [Errno 2] No such file or directory:

+++ meta/classes/kernel-module-split.bbclass
@@ -197,7 +197,7 @@ python split_kernel_module_packages () {
     # directory cannot be removed.
     dvar = d.getVar('PKGD', True)
     for dir in ["%s/etc/modprobe.d" % (dvar), "%s/etc/modules-load.d" % (dvar), "%s/etc" % (dvar)]:
-        if len(os.listdir(dir)) == 0:
+        if os.path.isdir(dir) and len(os.listdir(dir)) == 0:
             os.rmdir(dir)
 }
